api/requirements.txt
beautifulsoup4==4.13.3
fastapi==0.115.9
httpx==0.28.1
lxml==5.3.1
openai>=1.0.0
playwright==1.51.0
pydantic==2.11.2
pydantic-settings==2.8.1
python-dotenv>=1.0.0
tenacity==9.0.0
uvicorn==0.34.0
